requirements.txt
Django>=5.0,<6.1
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
